requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
streamlit>=1.35.0
jupyter
ipykernel

# Milestone 2 (agentic coach)
requests
feedparser
huggingface_hub
transformers
torch
fpdf2
